custom exceptions handling (WIP but more or less working)
I have added a lot of complexity with no clear benefit.
Using exceptions as a fallback mechanism may be enough for this
library while focusing on the features of the return based approach.
It would be nice, anyway, to add support for custom exceptions for
the soft failure case.
Even in the case of exceptions, the wrapped function is expected to
based on the return approach, if it runs succesfuly it must return
a map with at least :result and :value keys.